requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.6
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gprof2dot==2022.7.29
install==1.3.5
Pillow==9.4.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
